WIP - Add button to adder toolbar to show annotations for selection
Add a button to the adder toolbar to show the existing annotations that
relate to the selected text, along with a keyboard shortcut (the "v"
key) to trigger it.

This provides a way to filter annotations in the sidebar to those that
correspond to a particular highlight or region of text using only the
keyboard.